requirements_dev.txt
bumpversion>=0.5.3
coverage>=4.0
cryptography>=1.4
ecdsa>=0.13
flake8
Flask>=0.11.1
Flask-Cors>=2.1.2
futures>=3.0.5
Jinja2>=2.8
jsonnet>=0.8.9
jsonpatch>=1.14
pytest>=2.9.1
pytest-cov>=2.2.1
pytest-flask>=0.10.0
pytest-ordering
python-etcd>=0.4.3
PyYAML>=3.11
requests>=2.10.0
requests-mock
semantic-version>=2.5.0
Sphinx>=1.3.1
tabulate>=0.7.5
termcolor>=1.1.0
tox>=2.1.1
urllib3>=1.16
watchdog>=0.8.3
wheel>=0.23.0
sphinxcontrib-napoleon
-e git+https://github.com/ant31/pytest-sugar.git#egg=pytest-sugar
yapf
